+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -221,5 +221,5 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-10T13:44:11.298Z")
+    updated: Optional[datetime] = dateutil.parser.isoparse("2025-03-10T14:17:45.434Z")
     r"""The date and time the resource was last updated"""